+++ src/Service/DataSeedService.php
@@ -1,5 +1,8 @@
 <?php
 
+use ChurchCRM\Family;
+use ChurchCRM\Person;
+
 class DataSeedService
 {
   private function getPerson($rs, &$personPointer)
@@ -35,8 +38,19 @@ class DataSeedService
     for ($i = 0; $i < $families; $i++) {
       
       $hoh = $this->getPerson($rs, $personPointer);
-      
-      $FamilyID = $FamilyService->insertFamily($hoh);
+      $family = new Family();
+      $family->setName($hoh->name->last);
+      $family->setAddress1($hoh->location->street);
+      $family->setCity($hoh->location->city);
+      $family->setState($hoh->location->state);
+      $family->setCountry("USA");
+      $family->setZip($hoh->location->zip);
+      $family->setHomephone($hoh->phone);
+      $family->setDateEntered(date('Y-m-d h:i:s', $hoh->registered));
+      $family->setEnteredBy($_SESSION['iUserID']);
+      $family->save();
+      $FamilyID = $family->getId();
+
       $familyName = $hoh->name->last;
       $hoh->famID = $FamilyID;
       $hoh->per_fmr_id = 1;
@@ -46,10 +60,10 @@ class DataSeedService
       $spouse->famID = $FamilyID;
       $spouse->per_fmr_id = 2;
 
-      $hohID = $PersonService->insertPerson($hoh);
+      $hohID = $this->insertPerson($hoh);
       $this->savePersonImage($hoh, $hohID);
       $rTotalHoh += 1;
-      $spouseID = $PersonService->insertPerson($spouse);
+      $spouseID = $this->insertPerson($spouse);
       $this->savePersonImage($spouse, $spouseID);
       $rTotalSpouse += 1;
 
@@ -61,7 +75,7 @@ class DataSeedService
         $child->name->last = $familyName;
         $child->famID = $FamilyID;
         $child->per_fmr_id = 3;
-        $childID = $PersonService->insertPerson($child);
+        $childID = $this->insertPerson($child);
         $this->savePersonImage($child, $childID);
         $rTotalChildren += 1;
       }
@@ -94,5 +108,38 @@ class DataSeedService
     requireUserGroupMembership("bAdmin");
     echo '{"status":"Fundraisers Seed Data Not Implemented"}';
   }
+
+  function insertPerson($user)
+  {
+    $person = new Person();
+    $person->setTitle($user->name->title);
+    $person->setFirstName($user->name->first);
+    $person->setLastName($user->name->last);
+    if (FilterInput($user->gender) == "male") {
+      $person->setGender(1);
+    } else {
+      $person->setGender(2);
+    }
+
+    $person->setAddress1($user->location->street);
+    $person->setCity($user->location->city);
+    $person->setState($user->location->state);
+    $person->setZip($user->location->zip);
+    $person->setCountry("USA");
+    $person->setHomePhone($user->phone);
+    $person->setCellPhone($user->cell);
+    $person->setEmail($user->email);
+    $person->setBirthDay(date('d', $user->dob));
+    $person->setBirthMonth(date('m', $user->dob));
+    $person->setBirthYear(date('Y', $user->dob));
+    $person->setFamId($user->famID);
+    $person->setFmrId($user->per_fmr_id);
+    $person->setEnteredBy(FilterInput($_SESSION['iUserID']));
+    $person->setDateEntered(date('Y-m-d h:i:s', $user->registered));
+    $person->save();
+    return $person->getId();
+
+  }
+
 
 }
